Fix issues related to version 4.12.0 (#258)
* Limit minimum screen brightness.

* Check container existence before deletion in PasskeysManager

* Revert: Disable cache for SRAM3 used as DMA memory.

* Disable DMA for UART TX.

+++ core/src/trezor/lvglui/scrs/homescreen.py
@@ -604,8 +604,9 @@ class PasskeysManager(Screen):
             self._init = True
         else:
             if not self.is_visible():
-                self.container.delete()
-                del self.container
+                if hasattr(self, "container") and self.container:
+                    self.container.delete()
+                    del self.container
                 self.fresh_show()
                 lv.scr_load(self)
             return
